feat: new nexmark skeleton for distributed query processing
